2/8/23 wed
Now when you press the add button the input api added to an array and able to see on console. And when bin button pressed web site deletes the last api element and this can be seen on console. But it can't delete the api that pressed on the web site "yet".